requirements.txt
pyjwt
fastapi
uvicorn[standard]
pony
asyncio
Starlette